// src/backend/schemas/user.ts
export interface UserRequest {
    username: string;
    email: string;
    password: string;
}

export interface UserLogin {
    email: string;
    password: string;
}

export interface UserResponse {
    id: string;
    username: string;
    email: string;
    active: boolean;
    created_at: string;
    updated_at: string;
}